\name{cv.coef}
\alias{cv.coef}

\title{Coefficient of Variation}

\description{Compute coefficient of variation.}
\usage{
  cv.coef(x)
}
\arguments{
  \item{x}{A numeric vector.}
}

\value{
  A number within [0, 1].
}

\author{
Shengqiao Li<lishengqiao@yahoo.com>
}
\keyword{univar}