\name{lambda}
\alias{lambda}

\title{Compute Number of Silent Features}
\description{
Compute number of silent features for Random KNN
}
\usage{
lambda(p, m, r)
}
\arguments{
  \item{p}{Total number of available features.}  
  \item{m}{Number of features to be drawn for each KNN.}
  \item{r}{Number of KNN  to be generated.}
}

\value{ A scalar for the mean silent features. }

\author{
Shengqiao Li<lishengqiao@yahoo.com>
}
\keyword{distribution}
