\name{normalize}
\alias{normalize.decscale}
\alias{normalize.sigmoidal}
\alias{normalize.softmax}
\alias{normalize.unit}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Data Normalization}
\description{
Data matrix normalization procedures.
}
\usage{
normalize.decscale(data)
normalize.sigmoidal(data)
normalize.softmax(data)
normalize.unit(data)
}

\arguments{
  \item{data}{A data matrix to be normalized.}
}

\value{
A normalized data matrix.
}
 
\author{
Shengqiao Li<lishengqiao@yahoo.com>
} 
%\seealso{
%\code{\link[dprep]{decscale}},  \code{\link[dprep]{signorm}}, \code{\link[dprep]{softmaxnorm}} and \code{\link[dprep]{rangenorm}} in \code{dprep} package.
%} 
\keyword{manip}

