\name{rknn-package}
\alias{rknn-package}
\docType{package}
\title{Random KNN Classification and Regression}

\description{Random KNN Classification and Regression}

\details{
\tabular{ll}{
Package: \tab rknn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-22\cr
Depends: \tab R (>= 2.15.0), gmp, FNN\cr
Suggests: \tab Hmisc, Biobase, genefilter, golubEsets, chemometrics\cr
Imports: \tab class, FNN\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
Packaged: \tab 2012-04-2\cr
}

Index:
\preformatted{
PRESS                   Predicted Residual Sum of Squares
begKNN                  Backward Elimination Feature Selection with
                        Random KNN
bestset                 Extract the best subset of feature from
                        selection process
confusion               Classification Confusion Matrix and Accuracy
cv.coef                 Coefficient of Variation
eta                     Coverage Probability
fitted.randomKNN        Extract Model Fitted Values
knn.reg                 KNN Regression
knn.reg.cv              KNN Regression Cross-Validation
lambda                  Compute Number of Silent Features
normalize.decscale      Data Normalization
plot.begKNN             Plot Function for Recursive Backward
                        Elimination Feature Selection
plot.supportRKNN        Plot Function for Support Criterion
predicted               Prediced Value From a Linear Model
print.KNNregcv          Print Method for KNN Regression
                        Cross-validation
print.beKNN             Print Method for Recursive Backward Elimination
                        Feature Selection
print.randomKNN         Print method for Random KNN regression
                        cross-validation
print.supportRKNN       Print Method for Random KNN Support Criterion
r                       Choose number of KNNs
randomKNN               Random KNN Classification and Regression
rknn-package            Random KNN Classification and Regression
rsqp                    Predicted R-square
supportRKNN             Support Criterion
varUsed                 Features Used or Not Used in Random KNN
}
}

\references{                    
  Shengqiao Li, E James Harner and Donald A Adjeroh. 
  \emph{Random KNN feature selection - a fast and stable alternative to Random Forests.}
  BMC Bioinformatics 2011, 12:450. \url{http://www.biomedcentral.com/1471-2105/12/450}
}
\author{
Shengqiao Li

Maintainer: Shengqiao Li <shli@stat.wvu.edu>
}
\keyword{ package }

