\name{predicted}
\alias{predicted}

\title{Prediced Value From a Linear Model}
\description{Extraceted predicted values from a linear model.}
\usage{
predicted(obj)
}

\arguments{
  \item{obj}{A linear model.}
}

\value{
  A vector of ppredicted values.
}

\author{
Shengqiao Li<lishengqiao@yahoo.com>
}
\keyword{models}
