\name{PRESS}
\alias{PRESS}
\title{Predicted Residual Sum of Squares}
\description{Predicted Residual Sum of Squares}
\usage{
PRESS(obj)
}
\arguments{
  \item{obj}{A linear model or knn regression}
}

\value{
  returns predicted residual sum of squares
}

\author{
Shengqiao Li<shli@stat.wvu.edu>
}

\keyword{regression}
