\name{print.rknnBE}
\alias{print.rknnBE}
 
\title{Print Method for Recursive Backward Elimination Feature Selection}
\description{Print summary of  recursive backward elimination feature selection.}
\usage{
\method{print}{rknnBE}(x, ...)
}
\arguments{
  \item{x}{A \code{beKNN} object returned by \code{rknnBeg} or \code{rknnBel}.}
  \item{\dots}{Addtional arguments to \code{print} method.}
}


\author{
Shengqiao Li<lishengqiao@yahoo.com>
}

\keyword{print}
