\name{print.rknnSupport}
\alias{print.rknnSupport}

\title{Print Method for Random KNN Support Criterion}

\description{Print \code{support}  summary of the features.}

\usage{
\method{print}{rknnSupport}(x, ...)
}

\arguments{
  \item{x}{A list returned from \code{rknnSupport} or \code{rknnRegSupport}.}
  \item{\dots}{Additonal \code{print} arguments.}
}

\author{
Shengqiao Li<shli@stat.wvu.edu>
}

\keyword{print}
