\name{print.rknn}
\alias{print.rknn}

\title{Print method for Random KNN}

\description{Print method for Random KNN}

\usage{
\method{print}{rknn}(x, ...)
}

\arguments{
  \item{x}{A \code{rknn} object.}
  \item{\dots}{Additonal \code{print} arguments.}
}

\author{
Shengqiao Li<shli@stat.wvu.edu>
}

\keyword{print}
